refactor: move constants to advisor.constants and remove advisor.copy
